+++ packages/plugin-giphy/src/icons/InsertPluginIcon.jsx
@@ -4,14 +4,13 @@ import React from 'react';
 const InsertPluginIcon = props => (
   <svg
     xmlns="http://www.w3.org/2000/svg"
-    width={21}
-    height={14}
-    viewBox="0 0 21 14"
-    style={{ width: '21px' }}
+    width={19}
+    height={13}
+    viewBox="0 0 19 13"
     {...props}
   >
     <g fill="none" fillRule="evenodd" stroke="currentColor">
-      <path d="M11 1v13.067M15.51 14.099V1.472h4.502M15.154 6.141h3.85M7 3.437a2.35 2.35 0 0 0-2.318-1.965h-.598a2.59 2.59 0 0 0-2.59 2.59v6.803a2.753 2.753 0 1 0 5.506 0V9.001a1 1 0 0 0-1-1H3.804" />
+      <path d="M10 1v11.2M14.51 12.227V1.405h4.502M14.154 5.407h3.85M6 3.089a2.08 2.08 0 0 0-2.041-1.684h-.875a2.59 2.59 0 0 0-2.59 2.59v5.068a2.753 2.753 0 0 0 2.753 2.753h.346A2.407 2.407 0 0 0 6 9.409V8a1 1 0 0 0-1-1H2.804" />
     </g>
   </svg>
 );