+++ packages/common/src/Utils/urlValidators.js
@@ -5,7 +5,7 @@ const linkify = linkifyIt();
 const youtubeRegex = /(?:youtu\.be\/|youtube\.com\/(?:embed\/|v\/|watch\?v=|watch\?.+&v=))((\w|-){11})$/;
 const facebookRegex = /facebook\.com\/([^/?].+\/)?video(s|\.php)[/?].*$/;
 const vimeoRegex = /(?:www\.|player\.)?vimeo.com\/(?:(?:channels|ondemand)\/(?:\w+\/)?|groups\/([^/]*)\/videos\/|album\/(\d+)\/video\/|video\/|)(\d+)(?:$|\/|\?)/;
-const soundCloudRegex = /(?:www\.|m\.)?soundcloud\.com\/\S+\/\S+$/;
+const soundCloudRegex = /soundcloud\.com\/\S+\/\S+$/;
 /* eslint-enable max-len, no-useless-escape */
 
 const isYoutube = url => youtubeRegex.test(url);
@@ -15,6 +15,7 @@ const isVimeo = url => vimeoRegex.test(url);
 export const isVideoUrl = url => [isYoutube, isFacebook, isVimeo].some(f => f(url));
 
 export const isSoundCloudUrl = url => soundCloudRegex.test(url);
+export const matchSoundCloudUrl = url => url.match(soundCloudRegex);
 
 export const isValidUrl = url => linkify.test(url);