+++ examples/editor/src/PluginsConfig.js
@@ -63,6 +63,7 @@ export default {
   },
   [LINK_TYPE]: {
     onClick: (event, url) => console.log('link clicked!', url),
+    // autoLink: false
   },
   [CODE_BLOCK_TYPE]: {},
   [DIVIDER_TYPE]: {},