+++ CHANGELOG.md
@@ -10,6 +10,11 @@
 
 ## [Unreleased]
 
+### :bug: Bug Fix
+
+- `giphy`
+  - [#192](https://github.com/wix-incubator/rich-content/pull/192) removing container styles of giphy to keep original width to height ratio
+  
 <br/>
 
 ## 1.7.4 (Nov 15, 2018)